d done

+++ functions2.c
@@ -174,7 +174,7 @@ int			ft_atoi2(const char *str, char *end)
 	return (sign * res);
 }
 
-int digits_in_base(int value, int base)
+int digits_in_base(long long value, int base)
 {
 	int i;
 	i = 0;
